// client/src/context/login.admin.interfaces.tsx
export interface LoginResponse {
    usuario: CreadoPor;
    accessToken: string;
}
export interface CreadoPor {
    rol: string;
    estado: boolean;
    google: boolean;
    nombre: string;
    correo: string;
    uid: string;
    img?: string;
}


export interface LoginData {
    username: string;
    password: string;
}
export interface RegisterData {
    correo: string;
    password: string;
    nombre: string;
}

export interface ProductsResponse {
    total: number;
    productos: Product;
}

export interface Product {
    precio: number;
    _id: string;
    nombre: string;
    categoria: Categoria;
    usuario: Categoria;
    img?: string;
}

export interface Categoria {
    _id: string;
    nombre: string
}

// Generated by https://quicktype.io

export interface CategoriasResponse {
    total: number;
    categorias: Categoria[];
}

export interface Categoria {
    _id: string;
    nombre: string;
    usuario?: CreadoPor;
}

export interface CreadoPor {
    _id: string;
    nombre: string;
}


// Generated by https://quicktype.io

export interface CameraInterface {
    fileName: string;
    fileSize: number;
    height: number;
    type: string;
    uri: string;
    width: number;
}
export interface ErrorResponse {
    statusCode: number;
    message:    string;
    error:      string;
}